Create separate SQL package and enforce rules
