--- transactionoutbox-core/src/main/java/com/gruelbox/transactionoutbox/DialectFamily.java
@@ -1,9 +1,0 @@
-package com.gruelbox.transactionoutbox;
-
-
-@Beta
-public enum DialectFamily {
-  MY_SQL,
-  H2,
-  POSTGRESQL;
-}
