+++ transactionoutbox-core/src/main/java/com/gruelbox/transactionoutbox/jdbc/JdbcPersistor.java
@@ -1,10 +1,10 @@
 package com.gruelbox.transactionoutbox.jdbc;
 
-import com.gruelbox.transactionoutbox.Dialect;
 import com.gruelbox.transactionoutbox.Persistor;
 import com.gruelbox.transactionoutbox.PersistorWrapper;
-import com.gruelbox.transactionoutbox.SqlPersistor;
 import com.gruelbox.transactionoutbox.TransactionOutbox;
+import com.gruelbox.transactionoutbox.sql.Dialect;
+import com.gruelbox.transactionoutbox.sql.SqlPersistor;
 import java.sql.Connection;
 import java.util.concurrent.CompletableFuture;
