+++ transactionoutbox-core/src/main/java/com/gruelbox/transactionoutbox/sql/DialectFamily.java
@@ -1,0 +1,10 @@
+package com.gruelbox.transactionoutbox.sql;
+
+import com.gruelbox.transactionoutbox.Beta;
+
+@Beta
+public enum DialectFamily {
+  MY_SQL,
+  H2,
+  POSTGRESQL;
+}
