+++ transactionoutbox-core/src/test/java/com/gruelbox/transactionoutbox/sql/AbstractSqlPersistorTest.java
@@ -1,4 +1,4 @@
-package com.gruelbox.transactionoutbox;
+package com.gruelbox.transactionoutbox.sql;
 
 import static java.time.Instant.now;
 import static java.time.temporal.ChronoUnit.MILLIS;
@@ -15,6 +15,13 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.ea.async.Async;
+import com.gruelbox.transactionoutbox.AlreadyScheduledException;
+import com.gruelbox.transactionoutbox.Invocation;
+import com.gruelbox.transactionoutbox.OptimisticLockException;
+import com.gruelbox.transactionoutbox.Persistor;
+import com.gruelbox.transactionoutbox.Transaction;
+import com.gruelbox.transactionoutbox.TransactionManager;
+import com.gruelbox.transactionoutbox.TransactionOutboxEntry;
 import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
